+++ src/route_node.rs
@@ -1,13 +1,13 @@
 
 #[derive(Show, Copy, Encodable, Decodable, PartialEq, Eq)]
 pub struct RouteNode {
-  pub from: uint,
-  pub to: uint,
-  pub weight: uint,
+  pub from: usize,
+  pub to: usize,
+  pub weight: usize,
 }
 
 impl RouteNode {
-  pub fn new(f: uint, t: uint, w: uint) -> RouteNode {
+  pub fn new(f: usize, t: usize, w: usize) -> RouteNode {
     RouteNode{ from: f, to: t, weight: w }
   }
 }
